+++ bundle-polkadot-hw-ledger.js
@@ -9330,7 +9330,7 @@
 		hasRequiredPackageInfo = 1;
 		Object.defineProperty(packageInfo, "__esModule", { value: true });
 		packageInfo.packageInfo = void 0;
-		packageInfo.packageInfo = { name: '@polkadot/hw-ledger-transports', path: typeof __dirname === 'string' ? __dirname : 'auto', type: 'cjs', version: '13.2.1' };
+		packageInfo.packageInfo = { name: '@polkadot/hw-ledger-transports', path: typeof __dirname === 'string' ? __dirname : 'auto', type: 'cjs', version: '13.2.2' };
 		return packageInfo;
 	}
 
@@ -9397,7 +9397,9 @@
 	const genericLedgerApps = {
 	    bittensor: 'Bittensor',
 	    encointer: 'Encointer',
-	    integritee: 'Integritee'
+	    frequency: 'Frequency',
+	    integritee: 'Integritee',
+	    polimec: 'Polimec'
 	};
 	const ledgerApps = {
 	    ...prevLedgerRecord,