+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 13.2.2
 - extension-dapp 0.55.1
 - react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 3.11.1
 - phishing 0.24.1